Debian: T3641: fix wrong library path for x86_64
NOTE: This is only a workaround to get x86_64 builds running again but needs a
proper solution. Bug is also present in the latest upstream version [1].

Debian now comes with multiarch/multilib support so libraries are no longer
stored in /lib but rather /lib/x86_64-linux-gnu or any other architecture.

[1]: https://salsa.debian.org/live-team/live-boot/-/blob/debian/1%2520210208/backend/initramfs-tools/live.hook#L47